refactor: split up convoluted theme customizer context and component files
