+++ contexts/theme-customizer/index.ts
@@ -1,0 +1,5 @@
+export * from './ThemeCustomizerContext';
+export * from './types';
+export * from './useConfigUpdater';
+export * from './utils';
+export * from './hooks';
